requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
mistralai>=1.10.0
